Moving AuthStateListener to MainActivity
