+++ EasyFood/app/src/main/java/com/example/easyfood/view/MainActivity.java
@@ -4,9 +4,14 @@ import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
+import android.widget.Toast;
+
+import androidx.annotation.NonNull;
 
 import com.example.easyfood.R;
 import com.example.easyfood.view.customer.EateryActivity;
+import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.auth.FirebaseUser;
 
 /**
  * Main Activity
@@ -14,7 +19,7 @@ import com.example.easyfood.view.customer.EateryActivity;
  * TEMPORARY:
  * Contains a Login and a Register Button.
  */
-public class MainActivity extends BaseActivity {
+public class MainActivity extends BaseActivity implements FirebaseAuth.AuthStateListener {
     private Button loginButton;
     private Button registerButton;
 
@@ -28,6 +33,32 @@ public class MainActivity extends BaseActivity {
 
         setLoginButtonListener();
         setRegisterButtonListener();
+    }
+
+    @Override
+    protected void onStart() {
+        super.onStart();
+        firebaseAuth.addAuthStateListener(this);
+    }
+
+    @Override
+    protected void onStop() {
+        super.onStop();
+        firebaseAuth.removeAuthStateListener(this);
+    }
+
+    @Override
+    public void onAuthStateChanged(@NonNull FirebaseAuth firebaseAuth) {
+        FirebaseUser user = firebaseAuth.getCurrentUser();
+
+        if (user != null) {
+            Toast.makeText(getApplicationContext(), user.getEmail(), Toast.LENGTH_SHORT).show(); // DEBUGGING
+
+            Intent I = new Intent(getApplicationContext(), EateryActivity.class);
+            startActivity(I);
+        } else {
+            Toast.makeText(getApplicationContext(), "Login to continue", Toast.LENGTH_SHORT).show();
+        }
     }
 
     /**
